refactor continued: rtc -> pocketmage_clock.h capacative touch -> pocketmage_touch.h sysFunc -> pocketmage_sys.h pocketmage sys uses namespace pocketmage and various functions use sub namespaces. This can be altered to not use any namespace.
