+++ Code/PocketMage_V3/include/pocketmage.h
@@ -1,11 +1,12 @@
 #pragma once
 #include <globals.h>
+
 // Pocketmage library headers
 #include <pocketmage_eink.h>
 #include <pocketmage_oled.h>
 #include <pocketmage_sd.h>
 #include <pocketmage_kb.h>
 #include <pocketmage_bz.h>
-//#include <pocketmage_cap.h> // planned
-//#include <pocketmage_rtc.h> // planned
-//#include <pocketmage_sys.h> // planned
+#include <pocketmage_touch.h>
+#include <pocketmage_clock.h>
+#include <pocketmage_sys.h>
