+++ Code/PocketMage_V3/lib/pocketmage_eink/include/pocketmage_eink.h
@@ -36,7 +36,7 @@ using DisplayT = GxEPD2_BW<PanelT, PanelT::HEIGHT>;
 // ===================== EINK CLASS =====================
 class PocketmageEink {
 public:
-  explicit PocketmageEink(DisplayT& display);
+  explicit PocketmageEink(DisplayT& display) : display_(display) {}
 
   // Wire up external buffers/state used to read from globals
   void setTextBuffer(std::vector<String>* lines)                            { lines_ = lines; };               // reference to allLines
@@ -71,21 +71,21 @@ public:
 
 private:
   DisplayT&             display_; // class reference to hardware display object
-  bool                  forceSlowFullUpdate_ = false;
-  uint8_t               partialCounter_      = 0;
-  const GFXfont*        currentFont_         = nullptr;
-  uint8_t               fullRefreshAfter_    = FULL_REFRESH_AFTER;
+  bool                  forceSlowFullUpdate_  = false;
+  uint8_t               partialCounter_       = 0;
+  const GFXfont*        currentFont_          = nullptr;
+  uint8_t               fullRefreshAfter_     = FULL_REFRESH_AFTER;
 
-  std::vector<String>*  lines_               = nullptr;
-  String*               editingFile_         = nullptr;
+  std::vector<String>*  lines_                = nullptr;
+  String*               editingFile_          = nullptr;
 
-  volatile long*        dynamicScroll_       = nullptr;
+  volatile long*        dynamicScroll_        = nullptr;
 
   // font metrics
-  uint8_t               lineSpacing_         = 6;
-  uint8_t               maxCharsPerLine_     = 0;
-  uint8_t               maxLines_            = 0;
-  uint8_t               fontHeight_          = 0;
+  uint8_t               lineSpacing_          = 6;
+  uint8_t               maxCharsPerLine_      = 0;
+  uint8_t               maxLines_             = 0;
+  uint8_t               fontHeight_           = 0;
 };
 
 void wireEink();
