+++ Code/PocketMage_V3/lib/pocketmage_sd/include/pocketmage_sd.h
@@ -19,15 +19,15 @@ public:
   explicit PocketmageSD() {}
 
   // Wire up external buffers/state used to read from globals
-  void setFileSys(fs::FS* fileSys)      { fileSys_ = fileSys;}            // reference to fs::FS*
-  void setOled(PocketmageOled* oled)     { oled_ = oled;}                // reference to pocketmage oled object
-  void setOled(PocketmageOled& oled)     { oled_ = &oled;}                // overloaded reference to pocketmage oled object
-  void setEink(PocketmageEink* eink)     { eink_ = eink;}                // reference to pocketmage oled object
-  void setEink(PocketmageEink& eink)     { eink_ = &eink;}                // overloaded reference to pocketmage oled object
-  void setEditingFile(String* editingFile) { editingFile_ = editingFile;}        // reference to editingFile
-  void setFilesList(String* filesList)    {filesList_ = filesList;}                           // reference to filesList
-  void setNoSD(volatile bool* noSD)           {noSD_ = noSD;}                           // reference to noSD
-  void setNoTimeout(bool* noTimeout)         {noTimeout_ = noTimeout;}                 // reference to noTimeout
+  void setFileSys(fs::FS* fileSys)                 { fileSys_ = fileSys;}  // reference to fs::FS*
+  void setOled(PocketmageOled* oled)                     { oled_ = oled;}  // reference to pocketmage oled object
+  void setOled(PocketmageOled& oled)                    { oled_ = &oled;}  // overloaded reference to pocketmage oled object
+  void setEink(PocketmageEink* eink)                     { eink_ = eink;}  // reference to pocketmage eink object
+  void setEink(PocketmageEink& eink)                    { eink_ = &eink;}  // overloaded reference to pocketmage eink object
+  void setEditingFile(String* editingFile) { editingFile_ = editingFile;}  // reference to editingFile
+  void setFilesList(String* filesList)          {filesList_ = filesList;}  // reference to filesList
+  void setNoSD(volatile bool* noSD)                       {noSD_ = noSD;}  // reference to noSD
+  void setNoTimeout(bool* noTimeout)            {noTimeout_ = noTimeout;}  // reference to noTimeout
 
   // Main methods  To Do: remove arguments for fs::FS &fs and reference internal fs::FS* instead
   void listDir(fs::FS &fs, const char *dirname);
