+++ Code/PocketMage_V3/src/setup/buzzerSetup.cpp
@@ -1,14 +1,12 @@
 #include <pocketmage.h>
 
-
 // Initialization of bz class
-static PocketmageBZ bz(buzzer);
+static PocketmageBZ pm_bz(buzzer);
 
 // Setup for Buzzer Class
 void setupBZ() {
   BZ().playJingle(Jingles::Startup);
 }
-
 
 // Wire function  for Buzzer class
 // add any global references here + add set function to class header file
@@ -16,4 +14,4 @@ void wireBZ() {
 }
 
 // Access for other apps
-PocketmageBZ& BZ() { return bz; }
+PocketmageBZ& BZ() { return pm_bz; }
